Backend/requirements.txt
dotenv
chromadb
openai
google-generativeai
anthropic
mistralai
python-dotenv
transformers
sentence-transformers
requests
pandas
numpy
pymupdf
faiss-cpu
matplotlib
playwright
huggingface_hub
hf_xet
seaborn
plotly
umap-learn
python-multipart
streamlit
fastapi